multiple adjustments
