+++ examples/schema_registry.py
@@ -8,13 +8,13 @@ from confluent_kafka.schema_registry.avro import AvroDeserializer, SchemaRegistr
 schema_registry = SchemaRegistryClient(SCHEMA_REGISTRY_URL)
 
 
-def callback(payload):
-    print("HELLO CALLBACK")
+def consume(payload):
+    print("HELLO consume")
     payload["value"] = AvroDeserializer(schema_registry=schema_registry)(
         payload["value"]
     )
     return payload
 
 
-async def async_callback(payload):
-    return callback(payload=payload)
+async def async_consume(payload):
+    return consume(payload=payload)
